basic monorepo setup
