+++ packages/client/nuxt.config.ts
@@ -10,13 +10,8 @@ export default defineNuxtConfig({
   },
   runtimeConfig: {
     public: {
-      socketPort: 3001,
       url: process.env.URL || "http://localhost:3001",
     },
   },
   plugins: ["./plugins/socket.io"],
-  nitro: {
-    preset: "node-server",
-    plugins: ["./plugins/socket.io.server"],
-  },
 });
