+++ packages/client/plugins/socket.io.ts
@@ -1,5 +1,5 @@
 import { io, Socket } from "socket.io-client";
-import { ClientPayloads, ServerPayloads } from "utils/SocketEvent";
+import { ClientPayloads, ServerPayloads } from "../utils/SocketEvent";
 
 export default defineNuxtPlugin(() => {
   const config = useRuntimeConfig().public;
